+++ src/app/core/unit-types/neutrals/index.ts
@@ -1,3 +1,3 @@
 export * from './units';
 import './units';
-import './fire-elementals';
+import './elementals';
